+++ java/src/test/java/gherkin/formatter/model/StepTest.java
@@ -10,9 +10,9 @@ public class StepTest {
     @Test
     public void shouldProvideArgumentsForOutlineTokens() {
         Step step = new Step(Collections.<Comment>emptyList(), "Given ", "I have <number> cukes in <whose> belly", 10, null, null);
-        assertEquals(7, step.getOutlineArgs().get(0).getOffset());
+        assertEquals(7, step.getOutlineArgs().get(0).getOffset().intValue());
         assertEquals("<number>", step.getOutlineArgs().get(0).getVal());
-        assertEquals(25, step.getOutlineArgs().get(1).getOffset());
+        assertEquals(25, step.getOutlineArgs().get(1).getOffset().intValue());
         assertEquals("<whose>", step.getOutlineArgs().get(1).getVal());
     }
 }
